# Resize cached solution instead of re-allocating every solve