First improvement of what we discussed.

## SDDP benchmark

This branch:
```
julia> include("/Users/oscar/Desktop/xpress_benchmark.jl")
--------------------------------------------------------------------------------
                      SDDP.jl (c) Oscar Dowson, 2017-20

Numerical stability report
  Non-zero Matrix range     [1e+00, 1e+00]
  Non-zero Objective range  [1e-02, 6e+03]
  Non-zero Bounds range     [7e-01, 2e+05]
  Non-zero RHS range        [7e+03, 5e+04]
No problems detected

Solver: serial mode

 Iteration    Simulation       Bound         Time (s)    Proc. ID   # Solves
        1    1.950877e+09   2.450826e+05   2.596121e+01          1       9879
        2    8.848306e+08   4.947587e+05   3.834260e+01          1      19758
        3    8.358749e+08   1.064414e+06   5.070363e+01          1      29637
        4    5.263854e+08   3.137281e+06   6.342422e+01          1      39516
        5    4.270620e+08   3.790418e+06   7.590998e+01          1      49395
^C
Terminating training with status: interrupted
------------------------------------------------------------------------------
 ─────────────────────────────────────────────────────────────────────────────
                                      Time                   Allocations      
                              ──────────────────────   ───────────────────────
       Tot / % measured:           80.2s / 95.6%           15.3GiB / 93.6%    

 Section              ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────
 backward_pass             6    63.6s  82.9%   10.6s   12.1GiB  84.7%  2.02GiB
   solve_subproblem    49.6k    62.3s  81.3%  1.26ms   12.0GiB  84.1%   255KiB
 forward_pass              6    13.1s  17.1%   2.18s   2.19GiB  15.3%   374MiB
   solve_subproblem      720    13.0s  17.0%  18.1ms   2.19GiB  15.3%  3.11MiB
   sample_scenario         6   8.83ms  0.01%  1.47ms   1.57MiB  0.01%   268KiB
 calculate_bound           5   6.04ms  0.01%  1.21ms   1.22MiB  0.01%   251KiB
 ─────────────────────────────────────────────────────────────────────────────
```

Current master:
```
julia> include("/Users/oscar/Desktop/xpress_benchmark.jl")
[ Info: Precompiling SDDP [f4570300-c277-11e8-125c-4912f86ce65d]
[ Info: Precompiling Xpress [9e70acf3-d6c9-5be6-b5bd-4e2c73e3e054]
[ Info: Xpress: Found license file /Applications/FICO Xpress/xpressmp/bin/xpauth.xpr
[ Info: Xpress: Development license detected.
--------------------------------------------------------------------------------
                      SDDP.jl (c) Oscar Dowson, 2017-20

Numerical stability report
  Non-zero Matrix range     [1e+00, 1e+00]
  Non-zero Objective range  [1e-02, 6e+03]
  Non-zero Bounds range     [7e-01, 2e+05]
  Non-zero RHS range        [7e+03, 5e+04]
No problems detected

Solver: serial mode

 Iteration    Simulation       Bound         Time (s)    Proc. ID   # Solves
        1    2.025614e+09   2.450826e+05   2.613266e+01          1       9879
        2    8.920571e+08   5.102848e+05   3.865228e+01          1      19758
        3    8.408773e+08   2.423504e+06   5.125581e+01          1      29637
        4    6.387011e+08   3.319912e+06   6.387548e+01          1      39516
        5    4.197534e+08   5.168727e+06   7.659208e+01          1      49395
^C
Terminating training with status: interrupted
------------------------------------------------------------------------------
 ─────────────────────────────────────────────────────────────────────────────
                                      Time                   Allocations      
                              ──────────────────────   ───────────────────────
       Tot / % measured:           80.7s / 95.7%           15.4GiB / 93.7%    

 Section              ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────
 backward_pass             6    63.9s  82.8%   10.6s   12.2GiB  84.8%  2.04GiB
   solve_subproblem    49.4k    62.6s  81.1%  1.27ms   12.2GiB  84.2%   258KiB
 forward_pass              6    13.3s  17.2%   2.22s   2.19GiB  15.2%   374MiB
   solve_subproblem      720    13.2s  17.2%  18.4ms   2.19GiB  15.2%  3.12MiB
   sample_scenario         6   8.39ms  0.01%  1.40ms   1.56MiB  0.01%   266KiB
 calculate_bound           5   6.72ms  0.01%  1.34ms   1.24MiB  0.01%   254KiB
 ─────────────────────────────────────────────────────────────────────────────
```

For reference, Gurobi is:
```
julia> include("/Users/oscar/Desktop/xpress_benchmark.jl")
[ Info: Precompiling Gurobi [2e9cd046-0924-5485-92f1-d5272153d98b]
Academic license - for non-commercial use only
--------------------------------------------------------------------------------
                      SDDP.jl (c) Oscar Dowson, 2017-20

Numerical stability report
  Non-zero Matrix range     [1e+00, 1e+00]
  Non-zero Objective range  [1e-02, 6e+03]
  Non-zero Bounds range     [7e-01, 2e+05]
  Non-zero RHS range        [7e+03, 5e+04]
No problems detected

Solver: serial mode

 Iteration    Simulation       Bound         Time (s)    Proc. ID   # Solves
        1    1.619466e+09   7.601077e+05   4.986700e+00          1       9879
        2    9.200673e+08   1.442720e+06   6.737141e+00          1      19758
        3    9.625130e+08   1.486729e+06   8.583504e+00          1      29637
        4    6.768110e+08   2.609010e+07   1.041612e+01          1      39516
        5    4.849711e+08   2.639276e+07   1.220973e+01          1      49395
        6    4.769599e+08   2.685110e+07   1.400918e+01          1      59274
        7    4.081949e+08   6.901531e+07   1.579720e+01          1      69153
        8    5.397642e+08   6.901797e+07   1.760008e+01          1      79032
        9    3.706842e+08   6.902160e+07   1.941390e+01          1      88911
^C
Terminating training with status: interrupted
------------------------------------------------------------------------------
 ─────────────────────────────────────────────────────────────────────────────
                                      Time                   Allocations      
                              ──────────────────────   ───────────────────────
       Tot / % measured:           21.0s / 94.7%           4.04GiB / 80.2%    

 Section              ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────
 backward_pass            10    16.5s  83.3%   1.65s   2.79GiB  86.1%   286MiB
   solve_subproblem    90.0k    16.1s  81.1%   179μs   2.73GiB  84.4%  31.9KiB
 forward_pass             10    3.31s  16.7%   331ms    462MiB  13.9%  46.2MiB
   solve_subproblem    1.20k    3.30s  16.6%  2.75ms    458MiB  13.8%   391KiB
   sample_scenario        10   6.00ms  0.03%   600μs   2.39MiB  0.07%   245KiB
 calculate_bound           9   3.18ms  0.02%   354μs    368KiB  0.01%  40.9KiB
 ─────────────────────────────────────────────────────────────────────────────
```

I'm running some detailed profiling on my local machine. But it appears that the time-sink is not in `optimize!`, but in the setting the RHS terms. Argh.

cc @jd-lara @kdheepak