@joaquimg 2020-02-19
Time sink being RHS makes total sense.
I cache it locally and then pass all at once before solving.
Intercepting a vector modification and calling the vector method might help.