@joaquimg 2020-02-20
@odow can we merge?
Not a big difference in SDDP but the implementation looks clear and simple.